+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-- **osmdbt:** wrapper implementation ([#1](https://github.com/MapColonies/osmdbt-wrapper/issues/1)) ([5130859](https://github.com/MapColonies/osmdbt-wrapper/commit/5130859434b2b2745e2d13f94b22d5f8ede4085b))
+* **osmdbt:** wrapper implementation ([#1](https://github.com/MapColonies/osmdbt-wrapper/issues/1)) ([5130859](https://github.com/MapColonies/osmdbt-wrapper/commit/5130859434b2b2745e2d13f94b22d5f8ede4085b))
 
 # Changelog